+++ python-bindings/python/egglog/__init__.py
@@ -3,7 +3,14 @@ Package for creating e-graphs in Python.
 """
 
 from . import config, ipython_magic  # noqa: F401
-from .bindings import EggSmolError, StageInfo, TimeOnly, WithPlan  # noqa: F401
+from .bindings import (  # noqa: F401
+    EggSmolError,
+    StageInfo,
+    TimeOnly,
+    WithPlan,
+    fingerprint_ops_64,
+    fingerprint_ops_128,
+)
 from .builtins import *  # noqa: UP029
 from .conversion import *
 from .deconstruct import *